EntryViewHolder now has his own class
